requirements.txt
# Can be install by doing
# pip install -r requirements.txt
requests==2.2.0
putio.py==2.1.0
progressbar==2.2
sqlalchemy==0.9.1
flask==0.10.1
flask-restless==0.12.1
flask-restful==0.2.11
Flask-SQLAlchemy==1.0
